Refactor metadata
