+++ kifmm/src/fmm/field_translation/source_to_target/metadata.rs
@@ -1,7 +1,10 @@
 //! Implementation of traits to compute metadata for field translation operations.
 use std::collections::HashSet;
 
-use green_kernels::{traits::Kernel as KernelTrait, types::EvalType};
+use green_kernels::{
+    helmholtz_3d::Helmholtz3dKernel, laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait,
+    types::EvalType,
+};
 use itertools::Itertools;
 use num::{Float, Zero};
 use rlst::{
@@ -12,22 +15,34 @@ use rlst::{
 
 use crate::{
     fmm::{
+        constants::DEFAULT_SVD_THRESHOLD,
         field_translation::source_to_target::transfer_vector::compute_transfer_vectors,
-        helpers::{flip3, ncoeffs_kifmm},
-        types::{BlasFieldTranslation, BlasMetadata, FftFieldTranslation, FftMetadata},
+        helpers::{
+            coordinate_index_pointer, flip3, homogenous_kernel_scale, leaf_expansion_pointers,
+            leaf_scales, leaf_surfaces, level_expansion_pointers, level_index_pointer,
+            ncoeffs_kifmm, potential_pointers,
+        },
+        pinv::pinv,
+        types::{BlasFieldTranslation, BlasMetadata, Charges, FftFieldTranslation, FftMetadata},
+        KiFmm,
     },
     traits::{
         fftw::{Dft, DftType},
-        field::{ConfigureSourceToTargetData, SourceToTargetData},
+        field::{
+            SourceAndTargetTranslationMetadata, SourceToTargetData as SourceToTargetDataTrait,
+            SourcetoTargetTranslationMetadata,
+        },
+        fmm::{FmmKernel, FmmMetadata, SourceToTargetTranslation},
         general::{AsComplex, Epsilon},
-        tree::FmmTreeNode,
+        tree::{Domain as DomainTrait, FmmTree, FmmTreeNode, Tree},
     },
     tree::{
         constants::{
-            ALPHA_INNER, NCORNERS, NHALO, NSIBLINGS, NSIBLINGS_SQUARED, NTRANSFER_VECTORS_KIFMM,
+            ALPHA_INNER, ALPHA_OUTER, NCORNERS, NHALO, NSIBLINGS, NSIBLINGS_SQUARED,
+            NTRANSFER_VECTORS_KIFMM,
         },
         helpers::find_corners,
-        types::{Domain, MortonKey},
+        types::MortonKey,
     },
 };
 
@@ -42,21 +57,601 @@ fn find_cutoff_rank<T: Float + RlstScalar + Gemm>(singular_values: &[T], thresho
 
     singular_values.len() - 1
 }
-
-impl<Scalar, Kernel> FftFieldTranslation<Scalar, Kernel>
+impl<Scalar, SourceToTargetData> SourceAndTargetTranslationMetadata
+    for KiFmm<Scalar, Laplace3dKernel<Scalar>, SourceToTargetData>
 where
-    Scalar: RlstScalar + AsComplex + Dft + Default,
-    <Scalar as RlstScalar>::Real: RlstScalar + Default,
-    Kernel: KernelTrait<T = Scalar> + Default + Send + Sync,
+    Scalar: RlstScalar + Default + Epsilon,
+    Array<Scalar, BaseArray<Scalar, VectorContainer<Scalar>, 2>, 2>: MatrixSvd<Item = Scalar>,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default,
+    Self: SourceToTargetTranslation,
 {
-    /// Constructor for FFT based field translations
-    pub fn new() -> Self {
-        Self {
-            transfer_vectors: compute_transfer_vectors(),
-            ..Default::default()
+    fn source(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain;
+
+        // Compute required surfaces
+        let upward_equivalent_surface =
+            root.surface_grid(self.expansion_order, &domain, alpha_inner);
+        let upward_check_surface = root.surface_grid(self.expansion_order, &domain, alpha_outer);
+
+        let nequiv_surface = upward_equivalent_surface.len() / self.dim;
+        let ncheck_surface = upward_check_surface.len() / self.dim;
+
+        // Assemble matrix of kernel evaluations between upward check to equivalent, and downward check to equivalent matrices
+        // As well as estimating their inverses using SVD
+        let mut uc2e_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+        self.kernel.assemble_st(
+            EvalType::Value,
+            &upward_equivalent_surface[..],
+            &upward_check_surface[..],
+            uc2e_t.data_mut(),
+        );
+
+        // Need to transpose so that rows correspond to targets and columns to sources
+        let mut uc2e = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+        uc2e.fill_from(uc2e_t.transpose());
+
+        let (s, ut, v) = pinv(&uc2e, None, None).unwrap();
+
+        let mut mat_s = rlst_dynamic_array2!(Scalar, [s.len(), s.len()]);
+        for i in 0..s.len() {
+            mat_s[[i, i]] = Scalar::from_real(s[i]);
         }
+
+        let uc2e_inv_1 =
+            vec![empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view())];
+        let uc2e_inv_2 = vec![ut];
+
+        // Calculate M2M operator matrices
+        let children = root.children();
+        let mut m2m = vec![rlst_dynamic_array2!(
+            Scalar,
+            [nequiv_surface, 8 * nequiv_surface]
+        )];
+        let mut m2m_vec = vec![Vec::new()];
+
+        for (i, child) in children.iter().enumerate() {
+            let child_upward_equivalent_surface =
+                child.surface_grid(self.expansion_order, &domain, alpha_inner);
+
+            let mut pc2ce_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+
+            self.kernel.assemble_st(
+                EvalType::Value,
+                &child_upward_equivalent_surface,
+                &upward_check_surface,
+                pc2ce_t.data_mut(),
+            );
+
+            // Need to transpose so that rows correspond to targets, and columns to sources
+            let mut pc2ce = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+            pc2ce.fill_from(pc2ce_t.transpose());
+
+            let tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
+                uc2e_inv_1[0].view(),
+                empty_array::<Scalar, 2>()
+                    .simple_mult_into_resize(uc2e_inv_2[0].view(), pc2ce.view()),
+            );
+            let l = i * nequiv_surface * nequiv_surface;
+            let r = l + nequiv_surface * nequiv_surface;
+
+            m2m[0].data_mut()[l..r].copy_from_slice(tmp.data());
+            m2m_vec[0].push(tmp);
+        }
+
+        self.source = m2m;
+        self.source_vec = m2m_vec;
+        self.uc2e_inv_1 = uc2e_inv_1;
+        self.uc2e_inv_2 = uc2e_inv_2;
     }
 
+    fn target(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain;
+
+        // Compute required surfaces
+        let downward_equivalent_surface =
+            root.surface_grid(self.expansion_order, &domain, alpha_outer);
+        let downward_check_surface = root.surface_grid(self.expansion_order, &domain, alpha_inner);
+
+        let nequiv_surface = downward_equivalent_surface.len() / self.dim;
+        let ncheck_surface = downward_check_surface.len() / self.dim;
+
+        // Assemble matrix of kernel evaluations between upward check to equivalent, and downward check to equivalent matrices
+        // As well as estimating their inverses using SVD
+        let mut dc2e_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+        self.kernel.assemble_st(
+            EvalType::Value,
+            &downward_equivalent_surface[..],
+            &downward_check_surface[..],
+            dc2e_t.data_mut(),
+        );
+
+        // Need to transpose so that rows correspond to targets and columns to sources
+        let mut dc2e = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+        dc2e.fill_from(dc2e_t.transpose());
+
+        let (s, ut, v) = pinv::<Scalar>(&dc2e, None, None).unwrap();
+
+        let mut mat_s = rlst_dynamic_array2!(Scalar, [s.len(), s.len()]);
+        for i in 0..s.len() {
+            mat_s[[i, i]] = Scalar::from_real(s[i]);
+        }
+
+        let dc2e_inv_1 =
+            vec![empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view())];
+        let dc2e_inv_2 = vec![ut];
+
+        // Calculate M2M and L2L operator matrices
+        let children = root.children();
+        let mut l2l = vec![Vec::new()];
+
+        for (_i, child) in children.iter().enumerate() {
+            let child_downward_check_surface =
+                child.surface_grid(self.expansion_order, &domain, alpha_inner);
+            // Need to transpose so that rows correspond to targets, and columns to sources
+
+            let mut cc2pe_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+            self.kernel.assemble_st(
+                EvalType::Value,
+                &downward_equivalent_surface,
+                &child_downward_check_surface,
+                cc2pe_t.data_mut(),
+            );
+
+            // Need to transpose so that rows correspond to targets, and columns to sources
+            let mut cc2pe = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+            cc2pe.fill_from(cc2pe_t.transpose());
+            let mut tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
+                dc2e_inv_1[0].view(),
+                empty_array::<Scalar, 2>()
+                    .simple_mult_into_resize(dc2e_inv_2[0].view(), cc2pe.view()),
+            );
+            tmp.data_mut()
+                .iter_mut()
+                .for_each(|d| *d *= homogenous_kernel_scale(child.level()));
+
+            l2l[0].push(tmp);
+        }
+
+        self.target_vec = l2l;
+        self.dc2e_inv_1 = dc2e_inv_1;
+        self.dc2e_inv_2 = dc2e_inv_2;
+    }
+}
+
+impl<Scalar, SourceToTargetData> SourceAndTargetTranslationMetadata
+    for KiFmm<Scalar, Helmholtz3dKernel<Scalar>, SourceToTargetData>
+where
+    Scalar: RlstScalar<Complex = Scalar> + Default + Epsilon,
+    Array<Scalar, BaseArray<Scalar, VectorContainer<Scalar>, 2>, 2>: MatrixSvd<Item = Scalar>,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default,
+    Self: SourceToTargetTranslation,
+{
+    fn source(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain;
+
+        let depth = self.tree.source_tree().depth();
+
+        let mut curr = root;
+        let mut uc2e_inv_1 = Vec::new();
+        let mut uc2e_inv_2 = Vec::new();
+
+        // Calculate inverse upward check to equivalent matrices on each level
+        for _level in 0..=depth {
+            // Compute required surfaces
+            let upward_equivalent_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_inner);
+            let upward_check_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_outer);
+
+            let nequiv_surface = upward_equivalent_surface.len() / self.dim;
+            let ncheck_surface = upward_check_surface.len() / self.dim;
+
+            // Assemble matrix of kernel evaluations between upward check to equivalent, and downward check to equivalent matrices
+            // As well as estimating their inverses using SVD
+            let mut uc2e_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+            self.kernel.assemble_st(
+                EvalType::Value,
+                &upward_equivalent_surface[..],
+                &upward_check_surface[..],
+                uc2e_t.data_mut(),
+            );
+
+            // Need to transpose so that rows correspond to targets and columns to sources
+            let mut uc2e = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+            uc2e.fill_from(uc2e_t.transpose());
+
+            let (s, ut, v) = pinv(&uc2e, None, None).unwrap();
+
+            let mut mat_s = rlst_dynamic_array2!(Scalar, [s.len(), s.len()]);
+            for i in 0..s.len() {
+                mat_s[[i, i]] = Scalar::from_real(s[i]);
+            }
+
+            uc2e_inv_1
+                .push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view()));
+            uc2e_inv_2.push(ut);
+
+            curr = curr.first_child();
+        }
+
+        let mut curr = root;
+        let mut source = Vec::new();
+        let mut source_vec = Vec::new();
+
+        for level in 0..depth {
+            // Compute required surfaces
+            let upward_equivalent_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_inner);
+            let upward_check_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_outer);
+
+            let nequiv_surface = upward_equivalent_surface.len() / self.dim;
+            let ncheck_surface = upward_check_surface.len() / self.dim;
+
+            // Calculate M2M operator matrices on each level
+            let children = curr.children();
+            let mut m2m = rlst_dynamic_array2!(Scalar, [nequiv_surface, 8 * nequiv_surface]);
+            let mut m2m_vec = Vec::new();
+
+            for (i, child) in children.iter().enumerate() {
+                let child_upward_equivalent_surface =
+                    child.surface_grid(self.expansion_order, &domain, alpha_inner);
+
+                let mut pc2ce_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+
+                self.kernel.assemble_st(
+                    EvalType::Value,
+                    &child_upward_equivalent_surface,
+                    &upward_check_surface,
+                    pc2ce_t.data_mut(),
+                );
+
+                // Need to transpose so that rows correspond to targets, and columns to sources
+                let mut pc2ce = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+                pc2ce.fill_from(pc2ce_t.transpose());
+
+                let tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
+                    uc2e_inv_1[level as usize].view(),
+                    empty_array::<Scalar, 2>()
+                        .simple_mult_into_resize(uc2e_inv_2[level as usize].view(), pc2ce.view()),
+                );
+                let l = i * nequiv_surface * nequiv_surface;
+                let r = l + nequiv_surface * nequiv_surface;
+
+                m2m.data_mut()[l..r].copy_from_slice(tmp.data());
+                m2m_vec.push(tmp);
+            }
+
+            source.push(m2m);
+            source_vec.push(m2m_vec);
+            curr = curr.first_child();
+        }
+
+        self.uc2e_inv_1 = uc2e_inv_1;
+        self.uc2e_inv_2 = uc2e_inv_2;
+        self.source = source;
+        self.source_vec = source_vec;
+    }
+
+    fn target(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain;
+
+        let depth = self.tree.source_tree().depth();
+
+        let mut curr = root;
+        let mut dc2e_inv_1 = Vec::new();
+        let mut dc2e_inv_2 = Vec::new();
+
+        // Calculate inverse upward check to equivalent matrices on each level
+        for _level in 0..=depth {
+            // Compute required surfaces
+            let downward_equivalent_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_outer);
+            let downward_check_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_inner);
+
+            let nequiv_surface = downward_equivalent_surface.len() / self.dim;
+            let ncheck_surface = downward_check_surface.len() / self.dim;
+
+            // Assemble matrix of kernel evaluations between upward check to equivalent, and downward check to equivalent matrices
+            // As well as estimating their inverses using SVD
+            let mut dc2e_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+            self.kernel.assemble_st(
+                EvalType::Value,
+                &downward_equivalent_surface[..],
+                &downward_check_surface[..],
+                dc2e_t.data_mut(),
+            );
+
+            // Need to transpose so that rows correspond to targets and columns to sources
+            let mut dc2e = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+            dc2e.fill_from(dc2e_t.transpose());
+
+            let (s, ut, v) = pinv::<Scalar>(&dc2e, None, None).unwrap();
+
+            let mut mat_s = rlst_dynamic_array2!(Scalar, [s.len(), s.len()]);
+            for i in 0..s.len() {
+                mat_s[[i, i]] = Scalar::from_real(s[i]);
+            }
+
+            dc2e_inv_1
+                .push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view()));
+            dc2e_inv_2.push(ut);
+            curr = curr.first_child();
+        }
+
+        let mut curr = root;
+        let mut target = Vec::new();
+
+        for level in 0..depth {
+            // Compute required surfaces
+            let downward_equivalent_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_outer);
+            let downard_check_surface =
+                curr.surface_grid(self.expansion_order, &domain, alpha_inner);
+
+            let nequiv_surface = downward_equivalent_surface.len() / self.dim;
+            let ncheck_surface = downard_check_surface.len() / self.dim;
+
+            // Calculate M2M operator matrices on each level
+            let children = curr.children();
+            let mut l2l = Vec::new();
+
+            for (_i, child) in children.iter().enumerate() {
+                let child_downward_check_surface =
+                    child.surface_grid(self.expansion_order, &domain, alpha_inner);
+                // Need to transpose so that rows correspond to targets, and columns to sources
+
+                let mut cc2pe_t = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
+                self.kernel.assemble_st(
+                    EvalType::Value,
+                    &downward_equivalent_surface,
+                    &child_downward_check_surface,
+                    cc2pe_t.data_mut(),
+                );
+
+                // Need to transpose so that rows correspond to targets, and columns to sources
+                let mut cc2pe = rlst_dynamic_array2!(Scalar, [nequiv_surface, ncheck_surface]);
+                cc2pe.fill_from(cc2pe_t.transpose());
+                let tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
+                    dc2e_inv_1[(level + 1) as usize].view(),
+                    empty_array::<Scalar, 2>().simple_mult_into_resize(
+                        dc2e_inv_2[(level + 1) as usize].view(),
+                        cc2pe.view(),
+                    ),
+                );
+
+                l2l.push(tmp);
+            }
+
+            target.push(l2l);
+            curr = curr.first_child();
+        }
+
+        self.dc2e_inv_1 = dc2e_inv_1;
+        self.dc2e_inv_2 = dc2e_inv_2;
+
+        self.target_vec = target;
+    }
+}
+
+impl<Scalar> SourcetoTargetTranslationMetadata
+    for KiFmm<Scalar, Helmholtz3dKernel<Scalar>, BlasFieldTranslation<Scalar>>
+where
+    Scalar: RlstScalar<Complex = Scalar> + Default,
+    <Scalar as RlstScalar>::Real: Default,
+{
+    fn field_translation(&mut self) {}
+}
+
+impl<Scalar> SourcetoTargetTranslationMetadata
+    for KiFmm<Scalar, Helmholtz3dKernel<Scalar>, FftFieldTranslation<Scalar>>
+where
+    Scalar: RlstScalar<Complex = Scalar> + Default + AsComplex + Dft,
+    <Scalar as RlstScalar>::Real: RlstScalar + Default,
+{
+    fn field_translation(&mut self) {}
+}
+
+impl<Scalar> SourcetoTargetTranslationMetadata
+    for KiFmm<Scalar, Laplace3dKernel<Scalar>, BlasFieldTranslation<Scalar>>
+where
+    Scalar: RlstScalar + Default,
+    <Scalar as RlstScalar>::Real: Default,
+    Array<Scalar, BaseArray<Scalar, VectorContainer<Scalar>, 2>, 2>: MatrixSvd<Item = Scalar>,
+{
+    fn field_translation(&mut self) {
+        // Compute unique M2L interactions at Level 3 (smallest choice with all vectors)
+        // Compute interaction matrices between source and unique targets, defined by unique transfer vectors
+        let nrows = ncoeffs_kifmm(self.expansion_order);
+        let ncols = ncoeffs_kifmm(self.expansion_order);
+
+        let mut se2tc_fat = rlst_dynamic_array2!(Scalar, [nrows, ncols * NTRANSFER_VECTORS_KIFMM]);
+        let mut se2tc_thin = rlst_dynamic_array2!(Scalar, [nrows * NTRANSFER_VECTORS_KIFMM, ncols]);
+
+        let alpha = Scalar::from(ALPHA_INNER).unwrap().re();
+
+        for (i, t) in self.source_to_target.transfer_vectors.iter().enumerate() {
+            let source_equivalent_surface = t.source.surface_grid(
+                self.expansion_order,
+                &self.tree.source_tree().domain(),
+                alpha,
+            );
+            let nsources = source_equivalent_surface.len() / self.kernel.space_dimension();
+
+            let target_check_surface = t.target.surface_grid(
+                self.expansion_order,
+                &self.tree.source_tree().domain(),
+                alpha,
+            );
+            let ntargets = target_check_surface.len() / self.kernel.space_dimension();
+
+            let mut tmp_gram_t = rlst_dynamic_array2!(Scalar, [ntargets, nsources]);
+
+            self.kernel.assemble_st(
+                EvalType::Value,
+                &source_equivalent_surface[..],
+                &target_check_surface[..],
+                tmp_gram_t.data_mut(),
+            );
+
+            // Need to transpose so that rows correspond to targets, and columns to sources
+            let mut tmp_gram = rlst_dynamic_array2!(Scalar, [nsources, ntargets]);
+            tmp_gram.fill_from(tmp_gram_t.transpose());
+
+            let mut block = se2tc_fat
+                .view_mut()
+                .into_subview([0, i * ncols], [nrows, ncols]);
+            block.fill_from(tmp_gram.view());
+
+            let mut block_column = se2tc_thin
+                .view_mut()
+                .into_subview([i * nrows, 0], [nrows, ncols]);
+            block_column.fill_from(tmp_gram.view());
+        }
+
+        let mu = se2tc_fat.shape()[0];
+        let nvt = se2tc_fat.shape()[1];
+        let k = std::cmp::min(mu, nvt);
+
+        let mut u_big = rlst_dynamic_array2!(Scalar, [mu, k]);
+        let mut sigma = vec![Scalar::zero().re(); k];
+        let mut vt_big = rlst_dynamic_array2!(Scalar, [k, nvt]);
+
+        se2tc_fat
+            .into_svd_alloc(
+                u_big.view_mut(),
+                vt_big.view_mut(),
+                &mut sigma[..],
+                SvdMode::Reduced,
+            )
+            .unwrap();
+        let cutoff_rank = find_cutoff_rank(&sigma, self.source_to_target.threshold);
+        let mut u = rlst_dynamic_array2!(Scalar, [mu, cutoff_rank]);
+        let mut sigma_mat = rlst_dynamic_array2!(Scalar, [cutoff_rank, cutoff_rank]);
+        let mut vt = rlst_dynamic_array2!(Scalar, [cutoff_rank, nvt]);
+
+        u.fill_from(u_big.into_subview([0, 0], [mu, cutoff_rank]));
+        vt.fill_from(vt_big.into_subview([0, 0], [cutoff_rank, nvt]));
+        for (j, s) in sigma.iter().enumerate().take(cutoff_rank) {
+            unsafe {
+                *sigma_mat.get_unchecked_mut([j, j]) = Scalar::from(*s).unwrap();
+            }
+        }
+
+        // Store compressed M2L operators
+        let thin_nrows = se2tc_thin.shape()[0];
+        let nst = se2tc_thin.shape()[1];
+        let k = std::cmp::min(thin_nrows, nst);
+        let mut _gamma = rlst_dynamic_array2!(Scalar, [thin_nrows, k]);
+        let mut _r = vec![Scalar::zero().re(); k];
+        let mut st = rlst_dynamic_array2!(Scalar, [k, nst]);
+
+        se2tc_thin
+            .into_svd_alloc(
+                _gamma.view_mut(),
+                st.view_mut(),
+                &mut _r[..],
+                SvdMode::Reduced,
+            )
+            .unwrap();
+
+        let mut s_trunc = rlst_dynamic_array2!(Scalar, [nst, cutoff_rank]);
+        for j in 0..cutoff_rank {
+            for i in 0..nst {
+                unsafe { *s_trunc.get_unchecked_mut([i, j]) = *st.get_unchecked([j, i]) }
+            }
+        }
+
+        let mut c_u = Vec::new();
+        let mut c_vt = Vec::new();
+
+        for i in 0..self.source_to_target.transfer_vectors.len() {
+            let vt_block = vt.view().into_subview([0, i * ncols], [cutoff_rank, ncols]);
+
+            let tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
+                sigma_mat.view(),
+                empty_array::<Scalar, 2>().simple_mult_into_resize(vt_block.view(), s_trunc.view()),
+            );
+
+            let mut u_i = rlst_dynamic_array2!(Scalar, [cutoff_rank, cutoff_rank]);
+            let mut sigma_i = vec![Scalar::zero().re(); cutoff_rank];
+            let mut vt_i = rlst_dynamic_array2!(Scalar, [cutoff_rank, cutoff_rank]);
+
+            tmp.into_svd_alloc(u_i.view_mut(), vt_i.view_mut(), &mut sigma_i, SvdMode::Full)
+                .unwrap();
+
+            let directional_cutoff_rank =
+                find_cutoff_rank(&sigma_i, self.source_to_target.threshold);
+
+            let mut u_i_compressed =
+                rlst_dynamic_array2!(Scalar, [cutoff_rank, directional_cutoff_rank]);
+            let mut vt_i_compressed_ =
+                rlst_dynamic_array2!(Scalar, [directional_cutoff_rank, cutoff_rank]);
+
+            let mut sigma_mat_i_compressed =
+                rlst_dynamic_array2!(Scalar, [directional_cutoff_rank, directional_cutoff_rank]);
+
+            u_i_compressed
+                .fill_from(u_i.into_subview([0, 0], [cutoff_rank, directional_cutoff_rank]));
+            vt_i_compressed_
+                .fill_from(vt_i.into_subview([0, 0], [directional_cutoff_rank, cutoff_rank]));
+
+            for (j, s) in sigma_i.iter().enumerate().take(directional_cutoff_rank) {
+                unsafe {
+                    *sigma_mat_i_compressed.get_unchecked_mut([j, j]) = Scalar::from(*s).unwrap();
+                }
+            }
+
+            let vt_i_compressed = empty_array::<Scalar, 2>()
+                .simple_mult_into_resize(sigma_mat_i_compressed.view(), vt_i_compressed_.view());
+
+            c_u.push(u_i_compressed);
+            c_vt.push(vt_i_compressed);
+        }
+
+        let mut st_trunc = rlst_dynamic_array2!(Scalar, [cutoff_rank, nst]);
+        st_trunc.fill_from(s_trunc.transpose());
+
+        let result = BlasMetadata {
+            u,
+            st: st_trunc,
+            c_u,
+            c_vt,
+        };
+        self.source_to_target.metadata = result;
+        self.source_to_target.cutoff_rank = cutoff_rank;
+    }
+}
+
+impl<Scalar, Kernel> KiFmm<Scalar, Kernel, FftFieldTranslation<Scalar>>
+where
+    Scalar: RlstScalar + AsComplex + Default + Dft,
+    Kernel: KernelTrait<T = Scalar> + FmmKernel + Default + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default,
+{
     /// Computes the unique Green's function evaluations and places them on a convolution grid on the source box wrt to a given
     /// target point on the target box surface grid.
     ///
@@ -113,7 +708,7 @@ where
 
         let mut result = rlst_dynamic_array3!(Scalar, [npad, npad, npad]);
 
-        for (i, &j) in self.surf_to_conv_map.iter().enumerate() {
+        for (i, &j) in self.source_to_target.surf_to_conv_map.iter().enumerate() {
             result.data_mut()[j] = charges[i];
         }
 
@@ -180,54 +775,40 @@ where
     }
 }
 
-impl<Scalar, Kernel> SourceToTargetData for FftFieldTranslation<Scalar, Kernel>
-where
-    Scalar: RlstScalar + AsComplex + Default + Dft,
-    <Scalar as RlstScalar>::Real: RlstScalar + Default,
-    Kernel: KernelTrait<T = Scalar> + Default + Send + Sync,
-{
-    type Metadata = FftMetadata<<Scalar as AsComplex>::ComplexType>;
-}
-
-impl<Scalar, Kernel> ConfigureSourceToTargetData for FftFieldTranslation<Scalar, Kernel>
+impl<Scalar> SourcetoTargetTranslationMetadata
+    for KiFmm<Scalar, Laplace3dKernel<Scalar>, FftFieldTranslation<Scalar>>
 where
     Scalar: RlstScalar
         + AsComplex
         + Default
         + Dft<InputType = Scalar, OutputType = <Scalar as AsComplex>::ComplexType>,
     <Scalar as RlstScalar>::Real: RlstScalar + Default,
-    Kernel: KernelTrait<T = Scalar> + Default + Send + Sync,
 {
-    type Scalar = Scalar;
-    type Domain = Domain<Scalar::Real>;
-    type Kernel = Kernel;
-
-    fn expansion_order(&mut self, expansion_order: usize) {
-        self.expansion_order = expansion_order
-    }
-
-    fn kernel(&mut self, kernel: Self::Kernel) {
-        self.kernel = kernel
-    }
-
-    fn operator_data(&mut self, expansion_order: usize, domain: Self::Domain) {
-        // Parameters related to the FFT and Tree
-        let shape = <Self::Scalar as Dft>::shape_in(expansion_order);
-        let transform_shape = <Self::Scalar as Dft>::shape_out(expansion_order);
-        let transform_size = <Self::Scalar as Dft>::size_out(expansion_order);
+    fn field_translation(&mut self) {
+        // Compute the field translation operators
+        let shape = <Scalar as Dft>::shape_in(self.expansion_order);
+        let transform_shape = <Scalar as Dft>::shape_out(self.expansion_order);
+        let transform_size = <Scalar as Dft>::size_out(self.expansion_order);
 
         // Pick a point in the middle of the domain
         let two = Scalar::real(2.0);
-        let midway = domain.side_length.iter().map(|d| *d / two).collect_vec();
+        let midway = self
+            .tree
+            .source_tree()
+            .domain
+            .side_length
+            .iter()
+            .map(|d| *d / two)
+            .collect_vec();
         let point = midway
             .iter()
-            .zip(domain.origin)
-            .map(|(m, o)| *m + o)
+            .zip(self.tree.source_tree().domain().origin())
+            .map(|(m, o)| *m + *o)
             .collect_vec();
         let point = [point[0], point[1], point[2]];
 
         // Encode point in centre of domain and compute halo of parent, and their resp. children
-        let key = MortonKey::from_point(&point, &domain, 3);
+        let key = MortonKey::from_point(&point, &self.tree.source_tree().domain(), 3);
         let siblings = key.siblings();
         let parent = key.parent();
         let halo = parent.neighbors();
@@ -271,9 +852,9 @@ where
             sources[i] = tmp_sources;
         }
 
-        let n_source_equivalent_surface = 6 * (expansion_order - 1).pow(2) + 2;
+        let n_source_equivalent_surface = 6 * (self.expansion_order - 1).pow(2) + 2;
         let n_target_check_surface = n_source_equivalent_surface;
-        let alpha = Self::Scalar::real(ALPHA_INNER);
+        let alpha = Scalar::real(ALPHA_INNER);
 
         // Iterate over each set of convolutions in the halo (26)
         for i in 0..transfer_vectors.len() {
@@ -283,9 +864,16 @@ where
                 let target = targets[i][j];
                 let source = sources[i][j];
 
-                let source_equivalent_surface =
-                    source.surface_grid(expansion_order, &domain, alpha);
-                let target_check_surface = target.surface_grid(expansion_order, &domain, alpha);
+                let source_equivalent_surface = source.surface_grid(
+                    self.expansion_order,
+                    &self.tree.source_tree().domain(),
+                    alpha,
+                );
+                let target_check_surface = target.surface_grid(
+                    self.expansion_order,
+                    &&self.tree.source_tree().domain(),
+                    alpha,
+                );
 
                 let v_list: HashSet<MortonKey<_>> = target
                     .parent()
@@ -306,8 +894,8 @@ where
                     ];
 
                     let (conv_grid, _) = source.convolution_grid(
-                        expansion_order,
-                        &domain,
+                        self.expansion_order,
+                        &&self.tree.source_tree().domain(),
                         alpha,
                         &conv_point_corner,
                         conv_point_corner_index,
@@ -323,40 +911,33 @@ where
 
                     // Compute Green's fct evaluations
                     let mut kernel = flip3(&self.evaluate_greens_fct_convolution_grid(
-                        expansion_order,
+                        self.expansion_order,
                         &conv_grid,
                         kernel_point,
                     ));
 
                     // Compute FFT of padded kernel
-                    let mut kernel_hat = rlst_dynamic_array3!(
-                        <Self::Scalar as DftType>::OutputType,
-                        transform_shape
-                    );
+                    let mut kernel_hat =
+                        rlst_dynamic_array3!(<Scalar as DftType>::OutputType, transform_shape);
 
-                    let _ =
-                        Self::Scalar::forward_dft(kernel.data_mut(), kernel_hat.data_mut(), &shape);
+                    let _ = Scalar::forward_dft(kernel.data_mut(), kernel_hat.data_mut(), &shape);
 
                     kernel_data_vec[i].push(kernel_hat);
                 } else {
                     // Fill with zeros when interaction doesn't exist
-                    let kernel_hat_zeros = rlst_dynamic_array3!(
-                        <Self::Scalar as DftType>::OutputType,
-                        transform_shape
-                    );
+                    let kernel_hat_zeros =
+                        rlst_dynamic_array3!(<Scalar as DftType>::OutputType, transform_shape);
                     kernel_data_vec[i].push(kernel_hat_zeros);
                 }
             }
         }
 
         // Each element corresponds to all evaluations for each sibling (in order) at that halo position
-        let mut kernel_data = vec![
+        let mut kernel_data =
             vec![
-                <Self::Scalar as DftType>::OutputType::zero();
-                NSIBLINGS_SQUARED * transform_size
+                vec![<Scalar as DftType>::OutputType::zero(); NSIBLINGS_SQUARED * transform_size];
+                halo_children.len()
             ];
-            halo_children.len()
-        ];
 
         // For each halo position
         for i in 0..halo_children.len() {
@@ -396,10 +977,8 @@ where
                 let k_f =
                     &kernel_f[frequency_offset..(frequency_offset + NSIBLINGS_SQUARED)].to_vec();
                 let k_f_ = rlst_array_from_slice2!(k_f.as_slice(), [NSIBLINGS, NSIBLINGS]);
-                let mut k_ft = rlst_dynamic_array2!(
-                    <Self::Scalar as DftType>::OutputType,
-                    [NSIBLINGS, NSIBLINGS]
-                );
+                let mut k_ft =
+                    rlst_dynamic_array2!(<Scalar as DftType>::OutputType, [NSIBLINGS, NSIBLINGS]);
                 k_ft.fill_from(k_f_.view().transpose());
                 kernel_data_ft.push(k_ft.data().to_vec());
             }
@@ -411,26 +990,168 @@ where
         };
 
         // Set operator data
-        self.metadata = result;
+        self.source_to_target.metadata = result;
 
         // Set required maps, TODO: Should be a part of operator data
-        (self.surf_to_conv_map, self.conv_to_surf_map) =
-            FftFieldTranslation::<Self::Scalar, Kernel>::compute_surf_to_conv_map(
-                self.expansion_order,
-            );
+        (
+            self.source_to_target.surf_to_conv_map,
+            self.source_to_target.conv_to_surf_map,
+        ) = Self::compute_surf_to_conv_map(self.expansion_order);
+
+        // Set transfer vectors
+        self.source_to_target.transfer_vectors = compute_transfer_vectors();
     }
 }
 
-impl<Scalar, Kernel> BlasFieldTranslation<Scalar, Kernel>
+impl<Scalar, Kernel, SourceToTargetData> FmmMetadata for KiFmm<Scalar, Kernel, SourceToTargetData>
+where
+    Scalar: RlstScalar + Default,
+    Kernel: KernelTrait<T = Scalar> + FmmKernel + Default + Send + Sync,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default,
+{
+    // type Scalar = <Self as
+    type Scalar = Scalar;
+
+    fn metadata(&mut self, eval_type: EvalType, charges: &Charges<Self::Scalar>) {
+        let alpha_outer = Scalar::real(ALPHA_OUTER);
+
+        // Check if computing potentials, or potentials and derivatives
+        let eval_size = match eval_type {
+            EvalType::Value => 1,
+            EvalType::ValueDeriv => self.dim + 1,
+        };
+
+        // Check if we are computing matvec or matmul
+        let [_ncharges, nmatvecs] = charges.shape();
+        let ntarget_points = self.tree.target_tree.all_coordinates().unwrap().len() / self.dim;
+        let nsource_keys = self.tree.source_tree.n_keys_tot().unwrap();
+        let ntarget_keys = self.tree.target_tree.n_keys_tot().unwrap();
+        let ntarget_leaves = self.tree.target_tree.n_leaves().unwrap();
+        let nsource_leaves = self.tree.source_tree.n_leaves().unwrap();
+
+        // Buffers to store all multipole and local data
+        let multipoles = vec![Scalar::default(); self.ncoeffs * nsource_keys * nmatvecs];
+        let locals = vec![Scalar::default(); self.ncoeffs * ntarget_keys * nmatvecs];
+
+        // Index pointers of multipole and local data, indexed by level
+        let level_index_pointer_multipoles = level_index_pointer(&self.tree.source_tree);
+        let level_index_pointer_locals = level_index_pointer(&self.tree.target_tree);
+
+        // Buffer to store evaluated potentials and/or gradients at target points
+        let potentials = vec![Scalar::default(); ntarget_points * eval_size * nmatvecs];
+
+        // Kernel scale at each target and source leaf
+        let source_leaf_scales = leaf_scales::<Scalar>(
+            &self.tree.source_tree,
+            self.kernel.homogenous(),
+            self.ncoeffs,
+        );
+
+        // Pre compute check surfaces
+        let leaf_upward_surfaces_sources = leaf_surfaces(
+            &self.tree.source_tree,
+            self.ncoeffs,
+            alpha_outer,
+            self.expansion_order,
+        );
+        let leaf_upward_surfaces_targets = leaf_surfaces(
+            &self.tree.target_tree,
+            self.ncoeffs,
+            alpha_outer,
+            self.expansion_order,
+        );
+
+        // Mutable pointers to multipole and local data, indexed by level
+        let level_multipoles =
+            level_expansion_pointers(&self.tree.source_tree, self.ncoeffs, nmatvecs, &multipoles);
+
+        let level_locals =
+            level_expansion_pointers(&self.tree.source_tree, self.ncoeffs, nmatvecs, &locals);
+
+        // Mutable pointers to multipole and local data only at leaf level
+        let leaf_multipoles = leaf_expansion_pointers(
+            &self.tree.source_tree,
+            self.ncoeffs,
+            nmatvecs,
+            nsource_leaves,
+            &multipoles,
+        );
+
+        let leaf_locals = leaf_expansion_pointers(
+            &self.tree.target_tree,
+            self.ncoeffs,
+            nmatvecs,
+            ntarget_leaves,
+            &locals,
+        );
+
+        // Mutable pointers to potential data at each target leaf
+        let potentials_send_pointers = potential_pointers(
+            &self.tree.target_tree,
+            nmatvecs,
+            ntarget_leaves,
+            ntarget_points,
+            eval_size,
+            &potentials,
+        );
+
+        // Index pointer of charge data at each target leaf
+        let charge_index_pointer_targets = coordinate_index_pointer(&self.tree.target_tree);
+        let charge_index_pointer_sources = coordinate_index_pointer(&self.tree.source_tree);
+
+        // Set data
+        self.multipoles = multipoles;
+        self.locals = locals;
+        self.leaf_multipoles = leaf_multipoles;
+        self.level_multipoles = level_multipoles;
+        self.leaf_locals = leaf_locals;
+        self.level_locals = level_locals;
+        self.level_index_pointer_locals = level_index_pointer_locals;
+        self.level_index_pointer_multipoles = level_index_pointer_multipoles;
+        self.potentials = potentials;
+        self.potentials_send_pointers = potentials_send_pointers;
+        self.leaf_upward_surfaces_sources = leaf_upward_surfaces_sources;
+        self.leaf_upward_surfaces_targets = leaf_upward_surfaces_targets;
+        self.charges = charges.data().to_vec();
+        self.charge_index_pointer_targets = charge_index_pointer_targets;
+        self.charge_index_pointer_sources = charge_index_pointer_sources;
+        self.leaf_scales_sources = source_leaf_scales;
+        self.kernel_eval_size = eval_size;
+    }
+}
+
+impl<Scalar> FftFieldTranslation<Scalar>
+where
+    Scalar: RlstScalar + AsComplex + Dft + Default,
+    <Scalar as RlstScalar>::Real: RlstScalar + Default,
+{
+    /// Constructor for FFT based field translations
+    pub fn new() -> Self {
+        Self {
+            transfer_vectors: compute_transfer_vectors(),
+            ..Default::default()
+        }
+    }
+}
+
+impl<Scalar> SourceToTargetDataTrait for FftFieldTranslation<Scalar>
+where
+    Scalar: RlstScalar + AsComplex + Default + Dft,
+    <Scalar as RlstScalar>::Real: RlstScalar + Default,
+{
+    type Metadata = FftMetadata<<Scalar as AsComplex>::ComplexType>;
+}
+
+impl<Scalar> BlasFieldTranslation<Scalar>
 where
     Scalar: RlstScalar + Epsilon + Default,
     <Scalar as RlstScalar>::Real: Default,
-    Kernel: KernelTrait<T = Scalar> + Default,
 {
     /// Constructor for BLAS based field translations, specify a compression threshold for the SVD compressed operators
     /// TODO: More docs
     pub fn new(threshold: Option<Scalar::Real>) -> Self {
-        let tmp = Scalar::real(4) * Scalar::epsilon().re();
+        let tmp = Scalar::real(DEFAULT_SVD_THRESHOLD) * Scalar::epsilon().re();
 
         Self {
             threshold: threshold.unwrap_or(tmp),
@@ -440,233 +1161,61 @@ where
     }
 }
 
-impl<Scalar, Kernel> SourceToTargetData for BlasFieldTranslation<Scalar, Kernel>
+impl<Scalar> SourceToTargetDataTrait for BlasFieldTranslation<Scalar>
 where
     Scalar: RlstScalar,
-    Kernel: KernelTrait<T = Scalar> + Default,
 {
     type Metadata = BlasMetadata<Scalar>;
-}
-
-impl<Scalar, Kernel> ConfigureSourceToTargetData for BlasFieldTranslation<Scalar, Kernel>
-where
-    Scalar: RlstScalar,
-    Kernel: KernelTrait<T = Scalar> + Default,
-    Array<Scalar, BaseArray<Scalar, VectorContainer<Scalar>, 2>, 2>: MatrixSvd<Item = Scalar>,
-{
-    type Scalar = Scalar;
-    type Domain = Domain<Scalar::Real>;
-    type Kernel = Kernel;
-
-    fn expansion_order(&mut self, expansion_order: usize) {
-        self.expansion_order = expansion_order;
-    }
-
-    fn kernel(&mut self, kernel: Self::Kernel) {
-        self.kernel = kernel;
-    }
-
-    fn operator_data(&mut self, expansion_order: usize, domain: Self::Domain) {
-        // Compute unique M2L interactions at Level 3 (smallest choice with all vectors)
-        // Compute interaction matrices between source and unique targets, defined by unique transfer vectors
-        let nrows = ncoeffs_kifmm(expansion_order);
-        let ncols = ncoeffs_kifmm(expansion_order);
-
-        let mut se2tc_fat =
-            rlst_dynamic_array2!(Self::Scalar, [nrows, ncols * NTRANSFER_VECTORS_KIFMM]);
-        let mut se2tc_thin =
-            rlst_dynamic_array2!(Self::Scalar, [nrows * NTRANSFER_VECTORS_KIFMM, ncols]);
-
-        let alpha = Self::Scalar::from(ALPHA_INNER).unwrap().re();
-
-        for (i, t) in self.transfer_vectors.iter().enumerate() {
-            let source_equivalent_surface = t.source.surface_grid(expansion_order, &domain, alpha);
-            let nsources = source_equivalent_surface.len() / self.kernel.space_dimension();
-
-            let target_check_surface = t.target.surface_grid(expansion_order, &domain, alpha);
-            let ntargets = target_check_surface.len() / self.kernel.space_dimension();
-
-            let mut tmp_gram_t = rlst_dynamic_array2!(Self::Scalar, [ntargets, nsources]);
-
-            self.kernel.assemble_st(
-                EvalType::Value,
-                &source_equivalent_surface[..],
-                &target_check_surface[..],
-                tmp_gram_t.data_mut(),
-            );
-
-            // Need to transpose so that rows correspond to targets, and columns to sources
-            let mut tmp_gram = rlst_dynamic_array2!(Self::Scalar, [nsources, ntargets]);
-            tmp_gram.fill_from(tmp_gram_t.transpose());
-
-            let mut block = se2tc_fat
-                .view_mut()
-                .into_subview([0, i * ncols], [nrows, ncols]);
-            block.fill_from(tmp_gram.view());
-
-            let mut block_column = se2tc_thin
-                .view_mut()
-                .into_subview([i * nrows, 0], [nrows, ncols]);
-            block_column.fill_from(tmp_gram.view());
-        }
-
-        let mu = se2tc_fat.shape()[0];
-        let nvt = se2tc_fat.shape()[1];
-        let k = std::cmp::min(mu, nvt);
-
-        let mut u_big = rlst_dynamic_array2!(Self::Scalar, [mu, k]);
-        let mut sigma = vec![Self::Scalar::zero().re(); k];
-        let mut vt_big = rlst_dynamic_array2!(Self::Scalar, [k, nvt]);
-
-        se2tc_fat
-            .into_svd_alloc(
-                u_big.view_mut(),
-                vt_big.view_mut(),
-                &mut sigma[..],
-                SvdMode::Reduced,
-            )
-            .unwrap();
-        let cutoff_rank = find_cutoff_rank(&sigma, self.threshold);
-        let mut u = rlst_dynamic_array2!(Self::Scalar, [mu, cutoff_rank]);
-        let mut sigma_mat = rlst_dynamic_array2!(Self::Scalar, [cutoff_rank, cutoff_rank]);
-        let mut vt = rlst_dynamic_array2!(Self::Scalar, [cutoff_rank, nvt]);
-
-        u.fill_from(u_big.into_subview([0, 0], [mu, cutoff_rank]));
-        vt.fill_from(vt_big.into_subview([0, 0], [cutoff_rank, nvt]));
-        for (j, s) in sigma.iter().enumerate().take(cutoff_rank) {
-            unsafe {
-                *sigma_mat.get_unchecked_mut([j, j]) = Self::Scalar::from(*s).unwrap();
-            }
-        }
-
-        // Store compressed M2L operators
-        let thin_nrows = se2tc_thin.shape()[0];
-        let nst = se2tc_thin.shape()[1];
-        let k = std::cmp::min(thin_nrows, nst);
-        let mut _gamma = rlst_dynamic_array2!(Self::Scalar, [thin_nrows, k]);
-        let mut _r = vec![Self::Scalar::zero().re(); k];
-        let mut st = rlst_dynamic_array2!(Self::Scalar, [k, nst]);
-
-        se2tc_thin
-            .into_svd_alloc(
-                _gamma.view_mut(),
-                st.view_mut(),
-                &mut _r[..],
-                SvdMode::Reduced,
-            )
-            .unwrap();
-
-        let mut s_trunc = rlst_dynamic_array2!(Self::Scalar, [nst, cutoff_rank]);
-        for j in 0..cutoff_rank {
-            for i in 0..nst {
-                unsafe { *s_trunc.get_unchecked_mut([i, j]) = *st.get_unchecked([j, i]) }
-            }
-        }
-
-        let mut c_u = Vec::new();
-        let mut c_vt = Vec::new();
-
-        for i in 0..self.transfer_vectors.len() {
-            let vt_block = vt.view().into_subview([0, i * ncols], [cutoff_rank, ncols]);
-
-            let tmp = empty_array::<Self::Scalar, 2>().simple_mult_into_resize(
-                sigma_mat.view(),
-                empty_array::<Scalar, 2>().simple_mult_into_resize(vt_block.view(), s_trunc.view()),
-            );
-
-            let mut u_i = rlst_dynamic_array2!(Self::Scalar, [cutoff_rank, cutoff_rank]);
-            let mut sigma_i = vec![Self::Scalar::zero().re(); cutoff_rank];
-            let mut vt_i = rlst_dynamic_array2!(Self::Scalar, [cutoff_rank, cutoff_rank]);
-
-            tmp.into_svd_alloc(u_i.view_mut(), vt_i.view_mut(), &mut sigma_i, SvdMode::Full)
-                .unwrap();
-
-            let directional_cutoff_rank = find_cutoff_rank(&sigma_i, self.threshold);
-
-            let mut u_i_compressed =
-                rlst_dynamic_array2!(Self::Scalar, [cutoff_rank, directional_cutoff_rank]);
-            let mut vt_i_compressed_ =
-                rlst_dynamic_array2!(Self::Scalar, [directional_cutoff_rank, cutoff_rank]);
-
-            let mut sigma_mat_i_compressed = rlst_dynamic_array2!(
-                Self::Scalar,
-                [directional_cutoff_rank, directional_cutoff_rank]
-            );
-
-            u_i_compressed
-                .fill_from(u_i.into_subview([0, 0], [cutoff_rank, directional_cutoff_rank]));
-            vt_i_compressed_
-                .fill_from(vt_i.into_subview([0, 0], [directional_cutoff_rank, cutoff_rank]));
-
-            for (j, s) in sigma_i.iter().enumerate().take(directional_cutoff_rank) {
-                unsafe {
-                    *sigma_mat_i_compressed.get_unchecked_mut([j, j]) =
-                        Self::Scalar::from(*s).unwrap();
-                }
-            }
-
-            let vt_i_compressed = empty_array::<Self::Scalar, 2>()
-                .simple_mult_into_resize(sigma_mat_i_compressed.view(), vt_i_compressed_.view());
-
-            c_u.push(u_i_compressed);
-            c_vt.push(vt_i_compressed);
-        }
-
-        let mut st_trunc = rlst_dynamic_array2!(Self::Scalar, [cutoff_rank, nst]);
-        st_trunc.fill_from(s_trunc.transpose());
-
-        let result = BlasMetadata {
-            u,
-            st: st_trunc,
-            c_u,
-            c_vt,
-        };
-        self.metadata = result;
-        self.cutoff_rank = cutoff_rank;
-    }
 }
 
 #[cfg(test)]
 mod test {
     use green_kernels::laplace_3d::Laplace3dKernel;
     use num::Complex;
+    use rand::rngs::StdRng;
+    use rand::Rng;
+    use rand::SeedableRng;
     use rlst::RandomAccessByRef;
     use rlst::RandomAccessMut;
 
     use crate::fmm::helpers::flip3;
-    use crate::fmm::helpers::m2l_scale;
+    use crate::tree::helpers::points_fixture;
+    use crate::SingleNodeBuilder;
 
     use super::*;
 
     #[test]
-    fn test_blas_field_translation() {
-        let kernel = Laplace3dKernel::new();
+    fn test_blas_field_translation_laplace() {
+        // Setup random sources and targets
+        let nsources = 10000;
+        let ntargets = 10000;
+        let sources = points_fixture::<f64>(nsources, None, None, Some(1));
+        let targets = points_fixture::<f64>(ntargets, None, None, Some(1));
+
+        // FMM parameters
+        let n_crit = Some(100);
         let expansion_order = 6;
-        let domain = Domain::<f64>::new(&[0., 0., 0.], &[1., 1., 1.]);
-        let alpha = 1.05;
-        let threshold = 1e-5;
-        let cutoff_rank = 1000;
+        let sparse = true;
 
-        // Some expansion data
-        let ncoeffs = 6 * (expansion_order - 1usize).pow(2) + 2;
-        let mut multipole = rlst_dynamic_array2!(f64, [ncoeffs, 1]);
+        // Charge data
+        let nvecs = 1;
+        let mut rng = StdRng::seed_from_u64(0);
+        let mut charges = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        charges.data_mut().iter_mut().for_each(|c| *c = rng.gen());
 
-        for i in 0..ncoeffs {
-            *multipole.get_mut([i, 0]).unwrap() = i as f64;
-        }
-        let transfer_vectors = compute_transfer_vectors();
-
-        // Create field translation object
-        let mut blas = BlasFieldTranslation {
-            kernel,
-            threshold,
-            transfer_vectors,
-            expansion_order,
-            cutoff_rank,
-            ..Default::default()
-        };
-
-        blas.operator_data(expansion_order, domain);
+        let fmm = SingleNodeBuilder::new()
+            .tree(&sources, &targets, n_crit, sparse)
+            .unwrap()
+            .parameters(
+                &charges,
+                expansion_order,
+                Laplace3dKernel::new(),
+                EvalType::Value,
+                BlasFieldTranslation::new(Some(1e-5)),
+            )
+            .unwrap()
+            .build()
+            .unwrap();
 
         let idx = 123;
 
@@ -674,17 +1223,23 @@ mod test {
         let transfer_vector = &transfer_vectors[idx];
 
         // Lookup correct components of SVD compressed M2L operator matrix
-        let c_idx = blas
+        let c_idx = fmm
+            .source_to_target
             .transfer_vectors
             .iter()
             .position(|x| x.hash == transfer_vector.hash)
             .unwrap();
 
-        let c_u = &blas.metadata.c_u[c_idx];
-        let c_vt = &blas.metadata.c_vt[c_idx];
+        let c_u = &fmm.source_to_target.metadata.c_u[c_idx];
+        let c_vt = &fmm.source_to_target.metadata.c_vt[c_idx];
+
+        let mut multipole = rlst_dynamic_array2!(f64, [fmm.ncoeffs, 1]);
+        for i in 0..fmm.ncoeffs {
+            *multipole.get_mut([i, 0]).unwrap() = i as f64;
+        }
 
         let compressed_multipole = empty_array::<f64, 2>()
-            .simple_mult_into_resize(blas.metadata.st.view(), multipole.view());
+            .simple_mult_into_resize(fmm.source_to_target.metadata.st.view(), multipole.view());
 
         let compressed_check_potential = empty_array::<f64, 2>().simple_mult_into_resize(
             c_u.view(),
@@ -693,17 +1248,23 @@ mod test {
         );
 
         // Post process to find check potential
-        let check_potential = empty_array::<f64, 2>()
-            .simple_mult_into_resize(blas.metadata.u.view(), compressed_check_potential.view());
+        let check_potential = empty_array::<f64, 2>().simple_mult_into_resize(
+            fmm.source_to_target.metadata.u.view(),
+            compressed_check_potential.view(),
+        );
+
+        let alpha = ALPHA_INNER;
 
         let sources = transfer_vector
             .source
-            .surface_grid(expansion_order, &domain, alpha);
+            .surface_grid(expansion_order, &fmm.tree.domain, alpha);
         let targets = transfer_vector
             .target
-            .surface_grid(expansion_order, &domain, alpha);
-        let mut direct = vec![0f64; ncoeffs];
-        blas.kernel.evaluate_st(
+            .surface_grid(expansion_order, &fmm.tree.domain, alpha);
+
+        let mut direct = vec![0f64; fmm.ncoeffs];
+
+        fmm.kernel.evaluate_st(
             EvalType::Value,
             &sources[..],
             &targets[..],
@@ -720,124 +1281,6 @@ mod test {
         let rel_error: f64 = abs_error / (direct.iter().sum::<f64>());
 
         assert!(rel_error < 1e-5);
-    }
-
-    #[test]
-    fn test_fft_operator_data_kernels() {
-        let kernel = Laplace3dKernel::new();
-        let expansion_order: usize = 2;
-
-        let domain = Domain {
-            origin: [0., 0., 0.],
-            side_length: [1., 1., 1.],
-        };
-
-        // Some expansion data998
-        let ncoeffs = ncoeffs_kifmm(expansion_order);
-        let mut multipole = rlst_dynamic_array2!(f64, [ncoeffs, 1]);
-
-        for i in 0..ncoeffs {
-            *multipole.get_mut([i, 0]).unwrap() = i as f64;
-        }
-
-        let level = 2;
-
-        let transfer_vectors = compute_transfer_vectors();
-
-        // Create field translation object
-        let mut fft = FftFieldTranslation {
-            kernel,
-            expansion_order,
-            transfer_vectors,
-            ..Default::default()
-        };
-
-        fft.operator_data(expansion_order, domain);
-
-        let kernels = &fft.metadata.kernel_data;
-
-        let key = MortonKey::from_point(&[0.5, 0.5, 0.5], &domain, level);
-
-        let parent_neighbours = key.parent().neighbors();
-        let mut v_list_structured = vec![];
-        for _ in 0..26 {
-            v_list_structured.push(vec![]);
-        }
-        for (i, pn) in parent_neighbours.iter().enumerate() {
-            for child in pn.children() {
-                if !key.is_adjacent(&child) {
-                    v_list_structured[i].push(Some(child));
-                } else {
-                    v_list_structured[i].push(None)
-                }
-            }
-        }
-
-        // pick a halo position
-        let halo_idx = 0;
-        // pick a halo child position
-        let halo_child_idx = 2;
-        let n = 2 * expansion_order - 1;
-        let p = n + 1;
-        let size_real = p * p * (p / 2 + 1);
-
-        // Find kernel from precomputation;
-        let kernel_hat =
-            &kernels[halo_idx][halo_child_idx * size_real..(halo_child_idx + 1) * size_real];
-
-        // Apply scaling
-        let scale: f64 = m2l_scale(level).unwrap();
-        let kernel_hat = kernel_hat.iter().map(|k| *k * scale).collect_vec();
-
-        let target = key;
-        let source = v_list_structured[halo_idx][halo_child_idx].unwrap();
-        let source_equivalent_surface = source.surface_grid(expansion_order, &domain, ALPHA_INNER);
-        let target_check_surface = target.surface_grid(expansion_order, &domain, ALPHA_INNER);
-        let ntargets = target_check_surface.len() / 3;
-
-        // Compute conv grid
-        let conv_point_corner_index = 7;
-        let corners = find_corners(&source_equivalent_surface[..]);
-        let conv_point_corner = [
-            corners[conv_point_corner_index],
-            corners[8 + conv_point_corner_index],
-            corners[16 + conv_point_corner_index],
-        ];
-
-        let (conv_grid, _) = source.convolution_grid(
-            expansion_order,
-            &domain,
-            ALPHA_INNER,
-            &conv_point_corner,
-            conv_point_corner_index,
-        );
-
-        let kernel_point_index = 0;
-        let kernel_point = [
-            target_check_surface[kernel_point_index],
-            target_check_surface[ntargets + kernel_point_index],
-            target_check_surface[2 * ntargets + kernel_point_index],
-        ];
-
-        // Compute kernel from source/target pair
-        let test_kernel =
-            fft.evaluate_greens_fct_convolution_grid(expansion_order, &conv_grid, kernel_point);
-        let [m, n, o] = test_kernel.shape();
-
-        let mut test_kernel = flip3(&test_kernel);
-
-        // Compute FFT of padded kernel
-        let mut test_kernel_hat = rlst_dynamic_array3!(Complex<f64>, [m, n, o / 2 + 1]);
-        f64::forward_dft(
-            test_kernel.data_mut(),
-            test_kernel_hat.data_mut(),
-            &[m, n, o],
-        )
-        .unwrap();
-
-        for (p, t) in test_kernel_hat.data().iter().zip(kernel_hat.iter()) {
-            assert!((p - t).norm() < 1e-6)
-        }
     }
 
     #[test]
@@ -902,33 +1345,43 @@ mod test {
     }
 
     #[test]
-    fn test_fft_field_translation() {
-        let kernel = Laplace3dKernel::new();
-        let expansion_order: usize = 2;
+    fn test_fft_field_translation_laplace() {
+        // Setup random sources and targets
+        let nsources = 10000;
+        let ntargets = 10000;
+        let sources = points_fixture::<f64>(nsources, None, None, Some(1));
+        let targets = points_fixture::<f64>(ntargets, None, None, Some(1));
 
-        let domain = Domain {
-            origin: [0., 0., 0.],
-            side_length: [5., 5., 5.],
-        };
+        // FMM parameters
+        let n_crit = Some(100);
+        let expansion_order = 6;
+        let sparse = true;
 
-        let transfer_vectors = compute_transfer_vectors();
+        // Charge data
+        let nvecs = 1;
+        let mut rng = StdRng::seed_from_u64(0);
+        let mut charges = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        charges.data_mut().iter_mut().for_each(|c| *c = rng.gen());
 
-        // Some expansion data
-        let ncoeffs = ncoeffs_kifmm(expansion_order);
-        let mut multipole = rlst_dynamic_array2!(f64, [ncoeffs, 1]);
+        let fmm = SingleNodeBuilder::new()
+            .tree(&sources, &targets, n_crit, sparse)
+            .unwrap()
+            .parameters(
+                &charges,
+                expansion_order,
+                Laplace3dKernel::new(),
+                EvalType::Value,
+                FftFieldTranslation::new(),
+            )
+            .unwrap()
+            .build()
+            .unwrap();
 
-        for i in 0..ncoeffs {
+        let mut multipole = rlst_dynamic_array2!(f64, [fmm.ncoeffs, 1]);
+
+        for i in 0..fmm.ncoeffs {
             *multipole.get_mut([i, 0]).unwrap() = i as f64;
         }
-
-        let mut fft = FftFieldTranslation {
-            kernel,
-            expansion_order,
-            transfer_vectors,
-            ..Default::default()
-        };
-
-        fft.operator_data(expansion_order, domain);
 
         // Compute all M2L operators
         // Pick a random source/target pair
@@ -938,20 +1391,22 @@ mod test {
         let transfer_vector = &all_transfer_vectors[idx];
 
         // Compute FFT of the representative signal
-        let mut signal = fft.evaluate_charges_convolution_grid(expansion_order, multipole.data());
+        let mut signal = fmm.evaluate_charges_convolution_grid(expansion_order, multipole.data());
         let [m, n, o] = signal.shape();
         let mut signal_hat = rlst_dynamic_array3!(Complex<f64>, [m, n, o / 2 + 1]);
 
         let _ = f64::forward_dft(signal.data_mut(), signal_hat.data_mut(), &[m, n, o]);
 
-        let source_equivalent_surface =
-            transfer_vector
-                .source
-                .surface_grid(expansion_order, &domain, ALPHA_INNER);
-        let target_check_surface =
-            transfer_vector
-                .target
-                .surface_grid(expansion_order, &domain, ALPHA_INNER);
+        let source_equivalent_surface = transfer_vector.source.surface_grid(
+            expansion_order,
+            &fmm.tree.source_tree.domain,
+            ALPHA_INNER,
+        );
+        let target_check_surface = transfer_vector.target.surface_grid(
+            expansion_order,
+            &fmm.tree.source_tree.domain,
+            ALPHA_INNER,
+        );
         let ntargets = target_check_surface.len() / 3;
 
         // Compute conv grid
@@ -965,7 +1420,7 @@ mod test {
 
         let (conv_grid, _) = transfer_vector.source.convolution_grid(
             expansion_order,
-            &domain,
+            &fmm.tree.source_tree.domain,
             ALPHA_INNER,
             &conv_point_corner,
             conv_point_corner_index,
@@ -980,7 +1435,7 @@ mod test {
 
         // Compute kernel
         let kernel =
-            fft.evaluate_greens_fct_convolution_grid(expansion_order, &conv_grid, kernel_point);
+            fmm.evaluate_greens_fct_convolution_grid(expansion_order, &conv_grid, kernel_point);
         let [m, n, o] = kernel.shape();
 
         let mut kernel = flip3(&kernel);
@@ -1007,13 +1462,13 @@ mod test {
         );
 
         let mut result = vec![0f64; ntargets];
-        for (i, &idx) in fft.conv_to_surf_map.iter().enumerate() {
+        for (i, &idx) in fmm.source_to_target.conv_to_surf_map.iter().enumerate() {
             result[i] = potentials.data()[idx];
         }
 
         // Get direct evaluations for testing
-        let mut direct = vec![0f64; ncoeffs];
-        fft.kernel.evaluate_st(
+        let mut direct = vec![0f64; fmm.ncoeffs];
+        fmm.kernel.evaluate_st(
             EvalType::Value,
             &source_equivalent_surface[..],
             &target_check_surface[..],
